feat : 후기 삭제 API 구현

+++ src/main/java/dnd/project/domain/review/request/ReviewServiceRequest.java
@@ -16,7 +16,7 @@ public class ReviewServiceRequest {
     @Builder
     public static class Create {
         private Long lectureId;
-        private Integer score;
+        private Double score;
         private String tags;
         private Optional<String> content;
     }
